requirements.txt
# easyocr
fastapi
fastapi-cli
img2pdf
natasha
opencv_python
numpy
pdf2image
Pillow
pymystem3
pyocr
pytesseract
spacy
unoconv